add simulate option

+++ packages/create/src/cli/convert/eslint-config.ts
@@ -15,6 +15,7 @@ export function addConvertEslintCommand(cmd: Command): void {
 Examples:
     \`npx --yes @sap-ux/create@latest convert eslint-config\``
         )
+        .option('-s, --simulate', 'Simulate only. Do not write to the config file. Also, sets `--verbose`')
         .option('-v, --verbose', 'Show verbose information.')
         .option(
             '-c, --config <string>',
@@ -23,10 +24,10 @@ Examples:
         )
         .option('-n, --skip-install', 'Skip the `npm install` step.')
         .action(async (path, options) => {
-            if (options.verbose === true) {
+            if (options.verbose === true || options.simulate) {
                 setLogLevelVerbose();
             }
-            await convertEslintConfig(path || process.cwd(), options.config, !!options.skipInstall);
+            await convertEslintConfig(path || process.cwd(), !!options.simulate, options.config, !!options.skipInstall);
         });
 }
 
@@ -34,28 +35,36 @@ Examples:
  * Converts an eslint config to flat config format (eslint version 9).
  *
  * @param basePath - path to application root
+ * @param simulate - if true, simulates the conversion without writing to the config file. Also, sets `--verbose`
  * @param config - the name of the SAP Fiori tools eslint plugin config to be used
  * @param skipInstall - if true, skips the `npm install` step after converting the eslint config
  */
-async function convertEslintConfig(basePath: string, config: string, skipInstall = false): Promise<void> {
+async function convertEslintConfig(
+    basePath: string,
+    simulate: boolean,
+    config: string,
+    skipInstall = false
+): Promise<void> {
     const logger = getLogger();
     try {
-        logger.debug(`Called add eslint-config for path '${basePath}'`);
+        logger.debug(`Called add eslint-config for path '${basePath}', simulate is '${simulate}'`);
         await validateBasePath(basePath);
         const fs = await migrateEslintConfig(basePath, { logger, config });
         await traceChanges(fs);
-        fs.commit(() => {
-            logger.info(
-                `ESlint configuration converted. Ensure you install the new dependency by executing 'npm install'.`
-            );
-            if (skipInstall) {
+        if (!simulate) {
+            fs.commit(() => {
                 logger.info(
-                    `\`npm install\` was skipped. Ensure you install the dependencies before executing any linting commands.`
+                    `ESlint configuration converted. Ensure you install the new dependency by executing 'npm install'.`
                 );
-            } else {
-                runNpmInstallCommand(basePath);
-            }
-        });
+                if (skipInstall) {
+                    logger.info(
+                        `\`npm install\` was skipped. Ensure you install the dependencies before executing any linting commands.`
+                    );
+                } else {
+                    runNpmInstallCommand(basePath);
+                }
+            });
+        }
     } catch (error) {
         logger.error(`Error while executing convert eslint-config. '${(error as Error).message}'`);
         logger.debug(error as Error);
